Implement identity
Also fix a bug with `zeros` which resulting in incorrect array sizes
being returned, and refactor `offset` so it doesn't need a reference to
a `QArray`.

This is so that Rust doesn't complain when you pass a
reference to a `QArray` to a function (since it contains a `Vec` that
can't be safely shared). The correct way to do this is to pass slices,
but that doesn't happen automatically when you reference a struct
containing `Vec`s, so we'd need to implement a `as_ref` method or
something. For now, it was just simpler to pass slices to the individual
struct elements instead.

Fixes #6